[I25-170] refactor : 폴더 구조 변경
